+++ extensions/mozilla-firefox/src/hooks/useTabSearch.tsx
@@ -1,8 +1,8 @@
-import { useEffect, useState } from "react";
-import { Tab } from "../interfaces";
 import fs from "fs";
-import { getSessionActivePath, getSessionInactivePath } from "../util";
-import { decodeBlock as lz4DecodeBlock } from "../util";
+import { useEffect, useState } from "react";
+import { getPreferenceValues } from "@raycast/api";
+import { Preferences, Tab } from "../interfaces";
+import { decodeLZ4, getSessionActivePath, getSessionInactivePath } from "../util";
 
 const decompressSession = async (): Promise<any> => {
   let sessionPath = await getSessionInactivePath();
@@ -13,21 +13,11 @@ const decompressSession = async (): Promise<any> => {
     throw new Error("No profile session file found.");
   }
   const fileBuffer = await fs.promises.readFile(sessionPath);
-  const u8sz = fileBuffer.slice(8, 12);
-  const origLen = u8sz[0] + u8sz[1] * 256 + u8sz[2] * 256 * 256 + u8sz[3] * 256 * 256 * 256;
-  // Extract compressed data (past mozilla jsonlz4 header of 12 bytes)
-  const jsonStart = 12;
-  const u8Comp = fileBuffer.slice(jsonStart);
-  // Create LZ4 buffer
-  const comp = Buffer.from(u8Comp);
-  const orig = Buffer.alloc(origLen);
-  // perform lz4 decompression
-  const decompressedLen = lz4DecodeBlock(comp, orig);
-  const data = orig.subarray(0, decompressedLen);
-  return JSON.parse(data.toString());
+  return decodeLZ4(fileBuffer);
 };
 
 export function useTabSearch(query: string | undefined) {
+  const { tabSessionManagerId } = getPreferenceValues<Preferences>();
   const [entries, setEntries] = useState<Tab[]>();
   const [error, setError] = useState<string>();
   const [isLoading, setIsLoading] = useState<boolean>(true);
@@ -41,12 +31,17 @@ export function useTabSearch(query: string | undefined) {
       }
 
       try {
-        const session = await decompressSession();
-        let tabs: Tab[] = session.windows
-          .map((w: any) => w.tabs.map((t: any) => t.entries.map((e: any) => ({ url: e.url, title: e.title }))[0]))
-          .flat(2)
-          .filter((t: Tab) => t.url !== "about:newtab")
-          .map((e: any) => Tab.parse(e));
+        let tabs: Tab[] = [];
+
+        if (tabSessionManagerId) {
+        } else {
+          const session = await decompressSession();
+          tabs = session.windows
+            .map((w: any) => w.tabs.map((t: any) => t.entries.map((e: any) => ({ url: e.url, title: e.title }))[0]))
+            .flat(2)
+            .filter((t: Tab) => t.url !== "about:newtab")
+            .map((e: any) => Tab.parse(e));
+        }
 
         if (query) {
           tabs = tabs.filter((tab) => {
@@ -59,7 +54,6 @@ export function useTabSearch(query: string | undefined) {
 
         setEntries(tabs);
       } catch (e) {
-        console.error(e);
         setError((e as Error).message);
       } finally {
         if (!cancel) {