+++ extensions/mozilla-firefox/src/util/constants.ts
@@ -1,7 +1,0 @@
-export const SEARCH_ENGINE: { [key: string]: string } = {
-  google: `https://google.com/search?q=`,
-  bing: `https://www.bing.com/search?q=`,
-  baidu: `https://www.baidu.com/s?wd=`,
-  brave: `https://search.brave.com/search?q=`,
-  duckduckgo: `https://duckduckgo.com/?q=`,
-};